Improve UI from raw debug output to basic text UI. (#13)
